requirements.txt
# Core
fastapi
uvicorn

# LLM (Gemini)
google-generativeai

# Agent orchestration
fastmcp

# Utils
feedparser
requests
python-dotenv
pydantic

fastapi
uvicorn

